<b> install.sh run as root

+++ install.sh
@@ -4,6 +4,21 @@
 #<-[ nate-wilkins/dotfiles ]->#
 # \-------------------------/ #
 
+if [[ $(/usr/bin/id -u) -ne 0 ]]; then
+    echo "Unable to proceed. Script must be run as root."
+    echo "  sudo bash setup.sh"
+    exit -2
+fi
+
+if [[ "$1" == "" ]]; then
+    echo "Usage:                            "
+    echo "  sudo bash setup.sh <user-name>  "
+    echo "                                  "
+    echo "Example:                          "
+    echo "  sudo bash setup.sh nate-wilkins "
+    exit 1
+fi
+
 [ -d /home/$1/.fresh ]   && rm -rf /home/$1/.fresh
 [ -f /home/$1/.freshrc ] && rm -f  /home/$1/.freshrc
 find /home/$1/ -lname "/home/$1/.fresh/*" -delete
